+++ earth2studio/data/__init__.py
@@ -18,9 +18,11 @@ from .arco import ARCO
 from .base import DataSource, ForecastSource
 from .cbottle import CBottle3D
 from .cds import CDS
+from .cmip6 import CMIP6
 from .const import Constant, Constant_FX
 from .gefs import GEFS_FX, GEFS_FX_721x1440
 from .gfs import GFS, GFS_FX
+from .goes import GOES
 from .hrrr import HRRR, HRRR_FX
 from .ifs import IFS
 from .imerg import IMERG