+++ earth2studio/io/zarr.py
@@ -22,7 +22,6 @@ from typing import TYPE_CHECKING, Any, Union
 import numpy as np
 import torch
 import zarr
-from loguru import logger
 
 # Dealing with zarr 3.0 API breaks and type checking
 try:
@@ -36,13 +35,16 @@ if TYPE_CHECKING:
 
     from zarr.core import Array
     from zarr.core.array import Array as Array3
+    from zarr.core.array import CompressorsLike
 
     ZarrArray: TypeAlias = Union[Array, Array3]
 else:
     if zarr_major_version >= 3:
         ZarrArray = zarr.core.array.Array
+        from zarr.core.array import CompressorsLike
     else:
         ZarrArray = zarr.core.Array
+        CompressorsLike = Any
 
 from earth2studio.utils.coords import convert_multidim_to_singledim
 from earth2studio.utils.type import CoordSystem
@@ -62,10 +64,13 @@ class ZarrBackend:
         default {}
     backend_kwargs : dict[str, Any], optional
         Key word arguments for zarr.Group root object, by default {"overwrite": False}
+    zarr_codecs: CompressorsLike, optional
+        Compression codec to use when creating any new arrays. Only effects Zarr 3.0.
+        If None, will use no compressor, by default None
 
     Note
     ----
-    For keyword argument options see: https://zarr.readthedocs.io/en/stable/api/hierarchy.html
+    For keyword argument options see: https://zarr.readthedocs.io/en/latest/api/zarr/index.html#zarr.group
     """
 
     # sphinx - io zarr start
@@ -74,6 +79,7 @@ class ZarrBackend:
         file_name: str = None,
         chunks: dict[str, int] = {},
         backend_kwargs: dict[str, Any] = {"overwrite": False},
+        zarr_codecs: CompressorsLike = None,
     ) -> None:
 
         if file_name is None:
@@ -85,6 +91,7 @@ class ZarrBackend:
                 self.store = zarr.storage.DirectoryStore(file_name)
 
         self.root = zarr.group(self.store, **backend_kwargs)
+        self.zarr_codecs = zarr_codecs
 
         # Read data from file, if available
         self.coords: CoordSystem = OrderedDict({})
@@ -185,20 +192,9 @@ class ZarrBackend:
         for dim, values in adjusted_coords.items():
             if dim not in self.coords:
                 if zarr_major_version >= 3:
-                    # Dates types not supported in zarr 3.0 at the moment
-                    # https://github.com/zarr-developers/zarr-python/issues/2616
-                    # TODO: Remove once fixed
-                    if np.issubdtype(values.dtype, np.datetime64):
-                        logger.warning(
-                            "Datetime64 not supported in zarr 3.0, converting to int64 nanoseconds since epoch"
-                        )
-                        values = values.astype("datetime64[ns]").astype("int64")
 
-                    if np.issubdtype(values.dtype, np.timedelta64):
-                        logger.warning(
-                            "Timedelta64 not supported in zarr 3.0, converting to int64 nanoseconds since epoch"
-                        )
-                        values = values.astype("timedelta64[ns]").astype("int64")
+                    if "compressors" not in kwargs:
+                        kwargs["compressors"] = self.zarr_codecs
 
                     self.root.create_array(
                         dim,
@@ -226,20 +222,6 @@ class ZarrBackend:
             if k not in self.root:
                 values = coords[k]
                 if zarr_major_version >= 3:
-                    # Dates types not supported in zarr 3.0 at the moment
-                    # https://github.com/zarr-developers/zarr-python/issues/2616
-                    # TODO: Remove once fixed
-                    if np.issubdtype(values.dtype, np.datetime64):
-                        logger.warning(
-                            "Datetime64 not supported in zarr 3.0, converting to int64 nanoseconds since epoch"
-                        )
-                        values = values.astype("datetime64[ns]").astype("int64")
-
-                    if np.issubdtype(values.dtype, np.timedelta64):
-                        logger.warning(
-                            "Timedelta64 not supported in zarr 3.0, converting to int64 nanoseconds since epoch"
-                        )
-                        values = values.astype("timedelta64[ns]").astype("int64")
 
                     self.root.create_array(
                         k,